+++ plugin-gradle/src/main/java/com/diffplug/gradle/spotless/KotlinExtension.java
@@ -28,6 +28,7 @@ import org.gradle.api.tasks.SourceSet;
 
 import com.diffplug.spotless.FormatterStep;
 import com.diffplug.spotless.kotlin.KtLintStep;
+import com.diffplug.spotless.kotlin.KtfmtStep;
 
 public class KotlinExtension extends FormatExtension implements HasBuiltinDelimiterForLicense {
 	static final String NAME = "kotlin";
@@ -76,6 +77,33 @@ public class KotlinExtension extends FormatExtension implements HasBuiltinDelimi
 
 		private FormatterStep createStep() {
 			return KtLintStep.create(version, GradleProvisioner.fromProject(getProject()), userData);
+		}
+	}
+
+	/** Uses the [ktfmt](https://github.com/facebookincubator/ktfmt) jar to format source code. */
+	public KtfmtConfig ktfmt() {
+		return ktfmt(KtfmtStep.defaultVersion());
+	}
+
+	/**
+	 * Uses the given version of [ktfmt](https://github.com/facebookincubator/ktfmt) to format source
+	 * code.
+	 */
+	public KtfmtConfig ktfmt(String version) {
+		Objects.requireNonNull(version);
+		return new KtfmtConfig(version);
+	}
+
+	public class KtfmtConfig {
+		final String version;
+
+		KtfmtConfig(String version) {
+			this.version = Objects.requireNonNull(version);
+			addStep(createStep());
+		}
+
+		private FormatterStep createStep() {
+			return KtfmtStep.create(version, GradleProvisioner.fromProject(getProject()));
 		}
 	}